+++ rs/src/basic_data_types.rs
@@ -1,3 +1,4 @@
+use nom::number::complete::le_u64 as parse_le_u64;
 use nom::IResult as NomResult;
 
 /// Parse a null-terminated sequence of bytes. The nul-byte is consumed but not included in the
@@ -12,4 +13,11 @@ pub(crate) fn parse_c_string(input: &[u8]) -> NomResult<&[u8], String> {
     Ok(checked) => Ok((input, checked.to_string())),
     Err(_) => Err(nom::Err::Error((input, nom::error::ErrorKind::Verify))),
   }
+}
+
+pub(crate) fn parse_le32_f64(input: &[u8]) -> NomResult<&[u8], f64> {
+  let (input, bits) = parse_le_u64(input)?;
+  let bits = (bits >> 32) | (bits << 32);
+  let bytes = bits.to_le_bytes();
+  Ok((input, f64::from_le_bytes(bytes)))
 }