+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Next
+
+- **[Breaking change]** Update to `avm1-types@0.12.0`.
+- **[Feature]** Add support for the `StrictMode` action.
+- **[Fix]** Fix support for `float64` push value.
+
+## TypeScript
+
+- **[Internal]** Update to Yarn 2.
+
 # 0.11.0 (2020-09-07)
 
 - **[Breaking change]** Update to `avm1-types@0.11.0`.